requirements.txt
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
pytesseract
Pillow
python-dotenv
sentence-transformers
streamlit
